Base/App/Study/KVO/addFiles KVO (key-value Observing: キー値監視)パターン - KVOModel - KVOView
